refactor(token-stats): replace unpublished ledger with Room storage
Replace the spool, baseline, recovery, cleanup, and quarantine implementation with structured Room usage records and model settings. Integrate provider usage capture, SQL-backed statistics, settings and UI updates, and normal backup/restore handling. Document the redesigned storage model and remove obsolete tests.

+++ app/src/main/java/com/ai/assistance/operit/api/chat/llmprovider/LlamaProvider.kt
@@ -43,9 +43,9 @@ class LlamaProvider(
         }
     }
 
-    private var _inputTokenCount: Int = 0
-    private var _outputTokenCount: Int = 0
-    private var _cachedInputTokenCount: Int = 0
+    private var _inputTokenCount: Long = 0L
+    private var _outputTokenCount: Long = 0L
+    private var _cachedInputTokenCount: Long = 0L
 
     @Volatile
     private var isCancelled = false
@@ -54,21 +54,21 @@ class LlamaProvider(
     private var session: LlamaSession? = null
 
     override val inputTokenCount: Long
-        get() = _inputTokenCount.toLong()
+        get() = _inputTokenCount
 
     override val cachedInputTokenCount: Long
-        get() = _cachedInputTokenCount.toLong()
+        get() = _cachedInputTokenCount
 
     override val outputTokenCount: Long
-        get() = _outputTokenCount.toLong()
+        get() = _outputTokenCount
 
     override val providerModel: String
         get() = "${providerType.name}:$modelName"
 
     override fun resetTokenCounts() {
-        _inputTokenCount = 0
-        _outputTokenCount = 0
-        _cachedInputTokenCount = 0
+        _inputTokenCount = 0L
+        _outputTokenCount = 0L
+        _cachedInputTokenCount = 0L
     }
 
     private fun logLargeString(prefix: String, message: String) {
@@ -278,9 +278,9 @@ class LlamaProvider(
             }
         }
 
-        _inputTokenCount = kotlin.runCatching { s.countTokens(prompt) }.getOrElse { 0 }
-        _outputTokenCount = 0
-        onTokensUpdated(_inputTokenCount.toLong(), 0L, 0L)
+        _inputTokenCount = kotlin.runCatching { s.countTokens(prompt).toLong() }.getOrElse { 0L }
+        _outputTokenCount = 0L
+        onTokensUpdated(_inputTokenCount, 0L, 0L)
 
         val requestedMaxNewTokens = modelParameters
             .find { it.name == "max_tokens" }
@@ -292,7 +292,7 @@ class LlamaProvider(
             "开始llama.cpp推理，history=${chatHistory.size}, threads=${sessionConfig.nThreads}, n_ctx=${sessionConfig.nCtx}, n_batch=${sessionConfig.nBatch}, n_ubatch=${sessionConfig.nUBatch}, gpu_layers=${sessionConfig.nGpuLayers}, mmap=${sessionConfig.useMmap}"
         )
 
-        var outputTokenCount = 0
+        var outputTokenCount = 0L
         val toolCallOutputBuffer = StringBuilder()
         val finalOutputBuffer = StringBuilder()
 
@@ -306,7 +306,7 @@ class LlamaProvider(
                     if (isCancelled) {
                         false
                     } else {
-                        outputTokenCount += 1
+                        outputTokenCount += 1L
                         _outputTokenCount = outputTokenCount
 
                         if (effectiveEnableToolCall) {
@@ -318,7 +318,7 @@ class LlamaProvider(
 
                         kotlin.runCatching {
                             kotlinx.coroutines.runBlocking {
-                                onTokensUpdated(_inputTokenCount.toLong(), 0L, _outputTokenCount.toLong())
+                                onTokensUpdated(_inputTokenCount, 0L, _outputTokenCount)
                             }
                         }
